pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "be-conductor"
version = "0.3.26"
description = "Local orchestration layer for interactive terminal processes"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "click>=8.1.0",
    "httpx>=0.25.0",
    "qrcode>=7.0",
    "pyyaml>=6.0",
    "pyte>=0.8.0",
    "pywinpty>=2.0; sys_platform == 'win32'",
]

[tool.hatch.build.targets.wheel]
packages = ["be_conductor", "cli"]

[project.scripts]
be-conductor = "cli.main:cli"
